feat: adicionando usuários e autenticação JWT

+++ src/repositories/taskRepository.js
@@ -9,47 +9,47 @@ export default class TaskRepository {
             return task;
         
         } catch (error) {
-            return error;
+            throw error;
         }
     }
 
-    static async findAll() {
+    static async findAll(idUser) {
         try {
-            const tasks = await Task.findAll();
+            const tasks = await Task.findAll({where: {user_id: idUser}});
             return tasks;
         
         } catch (error) {
-            return error;
+            throw error;
         }
     }
 
-    static async FindById(id) {
+    static async FindById(id, idUser) {
         try {
-            const task = await Task.findByPk(id);
+            const task = await Task.findOne({where: {id: id, user_id: idUser}});
             return task;
         
         } catch (error) {
-            return error;
+            throw error;
         }
     }
 
-    static async update(id, data, transaction) {
+    static async update(id, idUser, data, transaction) {
         try {
-            const task = await Task.update(data, {where: {id: id}, transaction: transaction});
+            const task = await Task.update(data, {where: {id: id, user_id: idUser}, transaction: transaction});
             return task;
         
         } catch (error) {
-            return error;
+            throw error;
         }
     }
 
-    static async destroy(id, transaction) {
+    static async destroy(id, idUser, transaction) {
         try {
-            const task = await Task.destroy({where: {id: id}, transaction: transaction});
+            const task = await Task.destroy({where: {id: id, user_id: idUser}, transaction: transaction});
             return task;
         
         } catch (error) {
-            return error;
+            throw error;
         }
     }
 }
